backend/requirements.txt
flask==2.2.2
flask-cors==3.0.10
Flask-Session==0.4.0
flask-bcrypt==1.0.1
mysql-connector-python==8.0.31
Werkzeug==2.2.2
python-dotenv==1.0.0
google-cloud-vision==3.0.2
gunicorn==20.1.0